Resolve spinner issue in BucketFileList
Implemented same logic of data fetching and <Suspense> from BucketList
and only had to add fetcher inside the component. There seems to be no
issue with this on the client side. For future practice though, it may
be best to put the fetcher outside of the component definition when
possible